requirements.txt
Flask==3.1.2
Werkzeug==3.1.5
mysql-connector-python==8.0.33
gunicorn==21.2.0
flask-cors==6.0.2
